An implementation trying to provide the whole context to regular functions. It also changed the signature of the typeDiscoverCallback to be async.
Problem with code dublication. A lot of code gets almost dublicated. It's the same code but just adding asyn/sync. It's a mess to keep up.
Problem with Backwards compatibility. The aggregators need to be split because they call regular functions. They also call regular functions in there result function. Making it nessecary to make result async.
There are probably a lot more issues to be discovered within this code.
We should find a proper way t generalize the sync and async evaluators before continueing this implementation.